Improved the turbulence model

+++ hdr/header.hpp
@@ -1,8 +1,8 @@
 #ifndef FUNC_H_
 #define FUNC_H_
-int calculate_satellite_distance(double angle, int radius_earth, int radius_LEO);
-double integrate_turbulence(double angle, int radius_earth, int satellite_distance);
-int calculate_height(double angle, int radius_earth, int distance);
-float calculate_beam_spread(int satellite_distance, double integrated_turbulence, float diameter_laser, float wavenumber_laser, float curvature_laser, float inner_scale_size);
-float calculate_beam_intensity(int satellite_distance, double integrated_turbulence, float diameter_laser, float wavenumber_laser, float curvature_laser, float inner_scale_size);
+double calculate_satellite_distance(double angle, double radius_earth, double radius_LEO);
+double integrate_turbulence(double angle, double radius_earth, double satellite_distance);
+double calculate_height(double angle, double radius_earth, double distance);
+float calculate_beam_spread(double satellite_distance, double integrated_turbulence, double diameter_laser, double wavenumber_laser, double curvature_laser, float inner_scale_size);
+float calculate_beam_intensity(double satellite_distance, double integrated_turbulence, double diameter_laser, double wavenumber_laser, double curvature_laser, float inner_scale_size);
 #endif
